+++ internal/commands/versions_test.go
@@ -23,13 +23,13 @@ func TestVersionsCommand_ListVersions(t *testing.T) {
 			[]model.FoundVersion{
 				{
 					IndexVersion: model.IndexVersion{
-						TMID: "omnicorp/senseall/v0.36.0-20231231153548-243d1b462ccc.tm.json",
+						TMID: "murphy/omnicorp/senseall/v0.36.0-20231231153548-243d1b462ccc.tm.json",
 					},
 					FoundIn: model.FoundSource{RepoName: "r1"},
 				},
 				{
 					IndexVersion: model.IndexVersion{
-						TMID: "omnicorp/senseall/v0.35.0-20231230153548-243d1b462bbb.tm.json",
+						TMID: "murphy/omnicorp/senseall/v0.35.0-20231230153548-243d1b462bbb.tm.json",
 					},
 					FoundIn: model.FoundSource{RepoName: "r1"},
 				},
@@ -37,13 +37,13 @@ func TestVersionsCommand_ListVersions(t *testing.T) {
 		r2.On("Versions", mock.Anything, "senseall").Return([]model.FoundVersion{
 			{
 				IndexVersion: model.IndexVersion{
-					TMID: "omnicorp/senseall/v0.34.0-20231130153548-243d1b462aaa.tm.json",
+					TMID: "murphy/omnicorp/senseall/v0.34.0-20231130153548-243d1b462aaa.tm.json",
 				},
 				FoundIn: model.FoundSource{RepoName: "r2"},
 			},
 			{
 				IndexVersion: model.IndexVersion{
-					TMID: "omnicorp/senseall/v0.35.0-20231230173548-243d1b462bbb.tm.json",
+					TMID: "murphy/omnicorp/senseall/v0.35.0-20231230173548-243d1b462bbb.tm.json",
 				},
 				FoundIn: model.FoundSource{RepoName: "r2"},
 			},
@@ -56,15 +56,15 @@ func TestVersionsCommand_ListVersions(t *testing.T) {
 		assert.Len(t, res, 3)
 		assert.Equal(t, []model.FoundVersion{
 			{
-				IndexVersion: model.IndexVersion{TMID: "omnicorp/senseall/v0.34.0-20231130153548-243d1b462aaa.tm.json"},
+				IndexVersion: model.IndexVersion{TMID: "murphy/omnicorp/senseall/v0.34.0-20231130153548-243d1b462aaa.tm.json"},
 				FoundIn:      model.FoundSource{RepoName: "r2"},
 			},
 			{
-				IndexVersion: model.IndexVersion{TMID: "omnicorp/senseall/v0.35.0-20231230173548-243d1b462bbb.tm.json"},
+				IndexVersion: model.IndexVersion{TMID: "murphy/omnicorp/senseall/v0.35.0-20231230173548-243d1b462bbb.tm.json"},
 				FoundIn:      model.FoundSource{RepoName: "r2"},
 			},
 			{
-				IndexVersion: model.IndexVersion{TMID: "omnicorp/senseall/v0.36.0-20231231153548-243d1b462ccc.tm.json"},
+				IndexVersion: model.IndexVersion{TMID: "murphy/omnicorp/senseall/v0.36.0-20231231153548-243d1b462ccc.tm.json"},
 				FoundIn:      model.FoundSource{RepoName: "r1"},
 			},
 		}, res)
@@ -81,13 +81,13 @@ func TestVersionsCommand_ListVersions(t *testing.T) {
 			[]model.FoundVersion{
 				{
 					IndexVersion: model.IndexVersion{
-						TMID: "omnicorp/senseall/v0.36.0-20231231153548-243d1b462ccc.tm.json",
+						TMID: "murphy/omnicorp/senseall/v0.36.0-20231231153548-243d1b462ccc.tm.json",
 					},
 					FoundIn: model.FoundSource{RepoName: "r1"},
 				},
 				{
 					IndexVersion: model.IndexVersion{
-						TMID: "omnicorp/senseall/v0.35.0-20231230153548-243d1b462bbb.tm.json",
+						TMID: "murphy/omnicorp/senseall/v0.35.0-20231230153548-243d1b462bbb.tm.json",
 					},
 					FoundIn: model.FoundSource{RepoName: "r1"},
 				},
@@ -102,11 +102,11 @@ func TestVersionsCommand_ListVersions(t *testing.T) {
 		assert.Len(t, res, 2)
 		assert.Equal(t, []model.FoundVersion{
 			{
-				IndexVersion: model.IndexVersion{TMID: "omnicorp/senseall/v0.35.0-20231230153548-243d1b462bbb.tm.json"},
+				IndexVersion: model.IndexVersion{TMID: "murphy/omnicorp/senseall/v0.35.0-20231230153548-243d1b462bbb.tm.json"},
 				FoundIn:      model.FoundSource{RepoName: "r1"},
 			},
 			{
-				IndexVersion: model.IndexVersion{TMID: "omnicorp/senseall/v0.36.0-20231231153548-243d1b462ccc.tm.json"},
+				IndexVersion: model.IndexVersion{TMID: "murphy/omnicorp/senseall/v0.36.0-20231231153548-243d1b462ccc.tm.json"},
 				FoundIn:      model.FoundSource{RepoName: "r1"},
 			},
 		}, res)